app/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.34
python-dotenv==0.12.0
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.1
